move ChildProcessAppender and JsonLayout to org.transmart.logging
